Milestone 2 - seperate testing into JPMS-free test module
