+++ gorse-all/pom.xml
@@ -1,0 +1,30 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project xmlns="http://maven.apache.org/POM/4.0.0"
+         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+    <modelVersion>4.0.0</modelVersion>
+
+    <parent>
+        <groupId>io.gorse</groupId>
+        <artifactId>gorse-aggregator</artifactId>
+        <version>0.4.0</version>
+        <relativePath>../pom.xml</relativePath>
+    </parent>
+
+    <name>${project.groupId}:${project.artifactId}</name>
+    <artifactId>gorse-all</artifactId>
+    <description>Java SDK for Gorse recommender system.</description>
+
+    <dependencies>
+        <dependency>
+            <groupId>io.gorse</groupId>
+            <artifactId>gorse-client</artifactId>
+            <version>${project.version}</version>
+        </dependency>
+        <dependency>
+            <groupId>io.gorse</groupId>
+            <artifactId>gorse-test</artifactId>
+            <version>${project.version}</version>
+        </dependency>
+    </dependencies>
+</project>
